--- src/main/java/dto/VehicleDto.java
@@ -1,44 +1,0 @@
-package dto;
-
-public class VehicleDto extends BaseDto{
-
-    private String registrationID;
-
-    private String brand;
-
-    private String model;
-
-    private String bodyType;
-
-    public String getRegistrationID() {
-        return registrationID;
-    }
-
-    public void setRegistrationID(String registrationID) {
-        this.registrationID = registrationID;
-    }
-
-    public String getBrand() {
-        return brand;
-    }
-
-    public void setBrand(String brand) {
-        this.brand = brand;
-    }
-
-    public String getModel() {
-        return model;
-    }
-
-    public void setModel(String model) {
-        this.model = model;
-    }
-
-    public String getBodyType() {
-        return bodyType;
-    }
-
-    public void setBodyType(String bodyType) {
-        this.bodyType = bodyType;
-    }
-}
