+++ src/main/java/government/controller/VehiclesController.java
@@ -1,14 +1,13 @@
 package government.controller;
 
+import government.dto.VehicleDto;
 import government.facade.VehicleFacade;
 import government.model.Vehicle;
+import government.mapper.VehicleMapper;
 
 import javax.inject.Inject;
 import javax.transaction.Transactional;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
+import javax.ws.rs.*;
 import javax.ws.rs.core.Response;
 import java.util.Optional;
 
@@ -18,6 +17,9 @@ public class VehiclesController {
 
     @Inject
     VehicleFacade facade;
+
+    @Inject
+    VehicleMapper vehicleMapper;
 
     @GET
     @Path("{registration_id}")
@@ -30,5 +32,20 @@ public class VehiclesController {
         }
 
         return Response.ok(vehicle.get()).build();
+    }
+
+    @POST
+    @Transactional
+    public Response save(VehicleDto vehicleDto) {
+
+        Vehicle vehicle  = vehicleMapper.vehicleDtoToVehicle(vehicleDto);
+
+        vehicle = facade.save(vehicle);
+
+        if (vehicle.getId() == null){
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        }
+
+        return Response.status(Response.Status.CREATED).build();
     }
 }
