+++ src/main/java/government/dto/TripDto.java
@@ -1,4 +1,4 @@
-package dto;
+package government.dto;
 
 public class TripDto extends BaseDto {
