+++ src/main/java/government/dto/VehicleDto.java
@@ -1,20 +1,7 @@
-package model;
+package government.dto;
 
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import java.util.ArrayList;
-import java.util.List;
+public class VehicleDto extends BaseDto{
 
-@Entity
-@Table(name = "vehicles")
-public class Vehicle extends BaseEntity {
-
-    @NotNull
-    @NotBlank
     private String registrationID;
 
     private String brand;
@@ -22,9 +9,6 @@ public class Vehicle extends BaseEntity {
     private String model;
 
     private String bodyType;
-
-    @OneToMany(mappedBy = "vehicle", fetch = FetchType.LAZY)
-    private List<TrackerId> trackers = new ArrayList<>();
 
     public String getRegistrationID() {
         return registrationID;
@@ -56,13 +40,5 @@ public class Vehicle extends BaseEntity {
 
     public void setBodyType(String bodyType) {
         this.bodyType = bodyType;
-    }
-
-    public List<TrackerId> getTrackers() {
-        return trackers;
-    }
-
-    public void setTrackers(List<TrackerId> trackers) {
-        this.trackers = trackers;
     }
 }
