+++ src/main/java/government/model/Vehicle.java
@@ -1,9 +1,13 @@
 package government.model;
 
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @Table(name = "vehicles")
@@ -18,6 +22,9 @@ public class Vehicle extends BaseEntity {
     private String model;
 
     private String bodyType;
+
+    @OneToMany(mappedBy = "vehicle", fetch = FetchType.LAZY)
+    private List<TrackerId> trackers = new ArrayList<>();
 
     public String getRegistrationID() {
         return registrationID;
@@ -49,5 +56,13 @@ public class Vehicle extends BaseEntity {
 
     public void setBodyType(String bodyType) {
         this.bodyType = bodyType;
+    }
+
+    public List<TrackerId> getTrackers() {
+        return trackers;
+    }
+
+    public void setTrackers(List<TrackerId> trackers) {
+        this.trackers = trackers;
     }
 }
