+++ src/main/resources/META-INF/persistence.xml
@@ -9,7 +9,7 @@
         <jta-data-source>jdbc/government</jta-data-source>
         <properties>
             <property name="hibernate.dialect" value="org.hibernate.dialect.PostgresPlusDialect"/>
-            <property name="hibernate.hbm2ddl.auto" value="update"/>
+            <property name="hibernate.hbm2ddl.auto" value="create"/>
             <property name="hibernate.show_sql" value="true"/>
             <property name="hibernate.format_sql" value="true"/>
             <!--<property name="javax.persistence.schema-generation.database.action" value="drop-and-create" />-->
